requirements.txt
flask>=1.1.2
flask_restful>=0.3.8
flask_script==2.0.6
flask_migrate==2.1.1
marshmallow>=3.5.1
flask_sqlalchemy==2.3.2
flask_marshmallow>=0.8.0
marshmallow-sqlalchemy>=0.13.2
flask-restplus>=0.13.0
psycopg2>=2.7.5
numpy>=1.16.3
pymongo>=3.8.0
Werkzeug==0.16.1
more_itertools>=8.2.0
flask-mongoengine>=0.9.5
xlrd>=1.0.0
captcha>=0.3